LayoutTests/webaudio/decode-audio-data-basic-expected.txt

Basic tests for decodeAudioData function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.decodeAudioData(null, function(){}, function(){}); threw exception TypeError: Type error.
PASS The resources/media/24bit-44khz.wav test: successCallback has been called correctly.
PASS The resources/media/invalid-audio-file.txt test: errorCallback has been called correctly.
PASS successfullyParsed is true

TEST COMPLETE
